+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Support for P9813 RGB Led MOSFET controller (#10104)
 - Support for GPIO option selection
 - Gpio ``Option_a1`` enabling PWM2 high impedance if powered off as used by Wyze bulbs (#10196)
+- Support for FTC532 8-button touch controller by Peter Franck (#10222)
 
 ### Fixed
 - Redesign syslog and mqttlog using log buffer (#10164)